requirements.txt
torch
torchvision
torchaudio
librosa
numpy
matplotlib
pillow
streamlit